Fixes ruby 3.2 deprecation errors
Ruby 3.1 had warned of this problem, 3.2 totally deprecated it.

The issue was passing an integer cooresponding to the column type into
ActiveRecord, which tried to use a regular expression on it, causing the
error. I converted the value into a string to avoid the problem.

This is still a hack (as is the whole library) into the internals of
ActiveRecord, so other dragons could be there too.